# perf(GraphQL): misc enhancements