--- a/migrations/20241218093842-confirm-email-token-index.js
+++ b/migrations/20241218093842-confirm-email-token-index.js
@@ -1,0 +1,18 @@
+'use strict';
+
+/** @type {import('sequelize-cli').Migration} */
+module.exports = {
+  async up(queryInterface, Sequelize) {
+    await queryInterface.addIndex('Users', ['emailConfirmationToken'], {
+      unique: true,
+      where: {
+        deletedAt: { [Sequelize.Op.eq]: null },
+        emailConfirmationToken: { [Sequelize.Op.ne]: null },
+      },
+    });
+  },
+
+  async down(queryInterface, Sequelize) {
+    await queryInterface.removeIndex('Users', ['emailConfirmationToken']);
+  },
+};