## Gift cards query

Before
```
QUERY PLAN
Finalize GroupAggregate  (cost=45057.09..45059.30 rows=17 width=96) (actual time=6518.706..6532.104 rows=6 loops=1)
  Group Key: pm.batch
  ->  Gather Merge  (cost=45057.09..45059.22 rows=17 width=28) (actual time=6518.695..6532.091 rows=9 loops=1)
        Workers Planned: 1
        Workers Launched: 1
        ->  Partial GroupAggregate  (cost=44057.08..44057.41 rows=17 width=28) (actual time=6514.291..6514.309 rows=4 loops=2)
              Group Key: pm.batch
              ->  Sort  (cost=44057.08..44057.17 rows=181 width=24) (actual time=6514.283..6514.289 rows=92 loops=2)
                    Sort Key: pm.batch
                    Sort Method: quicksort  Memory: 26kB
                    Worker 0:  Sort Method: quicksort  Memory: 32kB
                    ->  Nested Loop  (cost=0.09..44055.73 rows=181 width=24) (actual time=37.274..6514.205 rows=92 loops=2)
"                          ->  Parallel Seq Scan on ""PaymentMethods"" pm  (cost=0.00..40586.92 rows=494306 width=28) (actual time=0.317..6252.498 rows=451232 loops=2)"
                          ->  Memoize  (cost=0.09..0.20 rows=1 width=4) (actual time=0.000..0.000 rows=0 loops=902465)
"                                Cache Key: pm.""SourcePaymentMethodId"""
                                Cache Mode: logical
                                Hits: 454355  Misses: 164  Evictions: 0  Overflows: 0  Memory Usage: 12kB
                                Worker 0:  Hits: 447776  Misses: 170  Evictions: 0  Overflows: 0  Memory Usage: 12kB
"                                ->  Index Scan using ""Cards_pkey"" on ""PaymentMethods"" spm  (cost=0.08..0.19 rows=1 width=4) (actual time=0.317..0.317 rows=0 loops=334)"
"                                      Index Cond: (id = pm.""SourcePaymentMethodId"")"
"                                      Filter: (""CollectiveId"" = 8686)"
                                      Rows Removed by Filter: 1
Planning Time: 0.650 ms
Execution Time: 6532.184 ms
```

After
```
QUERY PLAN
GroupAggregate  (cost=1507.24..1507.70 rows=17 width=96) (actual time=12.995..13.047 rows=5 loops=1)
  Group Key: pm.batch
  ->  Sort  (cost=1507.24..1507.37 rows=264 width=24) (actual time=12.981..12.998 rows=178 loops=1)
        Sort Key: pm.batch
        Sort Method: quicksort  Memory: 34kB
        ->  Merge Join  (cost=232.83..1505.12 rows=264 width=24) (actual time=0.363..12.936 rows=178 loops=1)
"              Merge Cond: (pm.""SourcePaymentMethodId"" = spm.id)"
"              ->  Index Scan using payment_methods__source_payment_method_id on ""PaymentMethods"" pm  (cost=0.08..68617.07 rows=778332 width=28) (actual time=0.035..8.153 rows=14014 loops=1)"
              ->  Sort  (cost=232.74..232.88 rows=285 width=4) (actual time=0.202..0.216 rows=150 loops=1)
                    Sort Key: spm.id
                    Sort Method: quicksort  Memory: 25kB
"                    ->  Index Scan using payment_methods__collective_id on ""PaymentMethods"" spm  (cost=0.08..230.42 rows=285 width=4) (actual time=0.036..0.183 rows=150 loops=1)"
"                          Index Cond: (""CollectiveId"" = 8686)"
Planning Time: 0.341 ms
Execution Time: 13.099 ms
```

## Email confirmation token mutation

Added index for slow `emailConfirmationToken` retrieval.

## Tier.totalDonated

Before
```
QUERY PLAN
Finalize GroupAggregate  (cost=1080.30..453927.04 rows=1 width=12) (actual time=1596.870..1629.494 rows=1 loops=1)
  ->  Gather  (cost=1080.30..453927.03 rows=1 width=12) (actual time=1590.024..1629.476 rows=2 loops=1)
        Workers Planned: 1
        Workers Launched: 1
        ->  Partial GroupAggregate  (cost=80.30..452926.93 rows=1 width=12) (actual time=1576.870..1576.872 rows=1 loops=2)
              ->  Nested Loop  (cost=80.30..452919.79 rows=14270 width=8) (actual time=55.891..1575.947 rows=4146 loops=2)
"                    ->  Parallel Bitmap Heap Scan on ""Orders"" ""Order""  (cost=80.21..21453.59 rows=8397 width=8) (actual time=54.331..152.653 rows=10941 loops=2)"
"                          Recheck Cond: ((""TierId"" = 3762) AND (""deletedAt"" IS NULL))"
                          Heap Blocks: exact=829
                          ->  Bitmap Index Scan on orders_tier_id  (cost=0.00..79.50 rows=14275 width=0) (actual time=31.198..31.199 rows=22094 loops=1)
"                                Index Cond: (""TierId"" = 3762)"
"                    ->  Index Scan using transactions__order_id on ""Transactions"" ""Transaction""  (cost=0.09..51.31 rows=23 width=8) (actual time=0.025..0.130 rows=0 loops=21882)"
"                          Index Cond: (""OrderId"" = ""Order"".id)"
"                          Filter: ((""RefundTransactionId"" IS NULL) AND (kind = ANY ('{CONTRIBUTION,ADDED_FUNDS}'::""enum_Transactions_kind""[])) AND ((type)::text = 'CREDIT'::text))"
                          Rows Removed by Filter: 2
Planning Time: 1.976 ms
JIT:
  Functions: 30
"  Options: Inlining false, Optimization false, Expressions true, Deforming true"
"  Timing: Generation 5.287 ms, Inlining 0.000 ms, Optimization 4.189 ms, Emission 68.407 ms, Total 77.883 ms"
Execution Time: 1673.242 ms
```

After
```
QUERY PLAN
Finalize GroupAggregate  (cost=1080.30..453926.94 rows=1 width=12) (actual time=169.859..183.685 rows=1 loops=1)
  ->  Gather  (cost=1080.30..453926.93 rows=1 width=12) (actual time=169.063..183.669 rows=2 loops=1)
        Workers Planned: 1
        Workers Launched: 1
        ->  Partial GroupAggregate  (cost=80.30..452926.83 rows=1 width=12) (actual time=135.251..135.253 rows=1 loops=2)
              ->  Nested Loop  (cost=80.30..452919.79 rows=14068 width=8) (actual time=34.267..134.319 rows=4146 loops=2)
"                    ->  Parallel Bitmap Heap Scan on ""Orders"" ""Order""  (cost=80.21..21453.59 rows=8397 width=8) (actual time=32.851..40.069 rows=10941 loops=2)"
"                          Recheck Cond: ((""TierId"" = 3762) AND (""deletedAt"" IS NULL))"
                          Heap Blocks: exact=1658
                          ->  Bitmap Index Scan on orders_tier_id  (cost=0.00..79.50 rows=14275 width=0) (actual time=1.352..1.352 rows=22094 loops=1)
"                                Index Cond: (""TierId"" = 3762)"
"                    ->  Index Scan using transactions__order_id on ""Transactions"" ""Transaction""  (cost=0.09..51.31 rows=23 width=8) (actual time=0.003..0.008 rows=0 loops=21882)"
"                          Index Cond: (""OrderId"" = ""Order"".id)"
"                          Filter: ((""RefundTransactionId"" IS NULL) AND (kind = ANY ('{CONTRIBUTION,ADDED_FUNDS}'::""enum_Transactions_kind""[])) AND ((type)::text = 'CREDIT'::text))"
                          Rows Removed by Filter: 2
Planning Time: 0.630 ms
JIT:
  Functions: 30
"  Options: Inlining false, Optimization false, Expressions true, Deforming true"
"  Timing: Generation 7.329 ms, Inlining 0.000 ms, Optimization 4.011 ms, Emission 58.916 ms, Total 70.257 ms"
Execution Time: 186.383 ms
```